+++ final/Web/service/src/main/java/com/expexchangeservice/service/interfaces/ILessonService.java
@@ -4,7 +4,6 @@ import com.expexchangeservice.model.dto.LessonDto;
 import com.expexchangeservice.model.dto.ProfileDto;
 import com.expexchangeservice.model.entities.Review;
 import com.expexchangeservice.model.entities.Theme;
-import com.expexchangeservice.model.entities.UserProfile;
 import com.expexchangeservice.model.enums.Type;
 
 import java.time.LocalDate;
@@ -14,17 +13,17 @@ import java.util.Set;
 public interface ILessonService {
 
     void addLesson(LessonDto lessonDto);
-    boolean changeLesson(int lessonId, LessonDto lessonDto);
-    boolean deleteLesson(Integer lessonId);
-    LessonDto getLessonById(Integer lessonId);
+    boolean changeLesson(Long lessonId, LessonDto lessonDto);
+    boolean deleteLesson(Long lessonId);
+    LessonDto getLessonById(Long lessonId);
     List<LessonDto> getAll();
     List<LessonDto> getLessonsOnTheDate(LocalDate date);
     List<LessonDto> getLessonsAfterDate(LocalDate date);
     List<LessonDto> getLessonsOnTheTheme(Theme theme);
     List<LessonDto> getLessonsForTheProfessor(ProfileDto profileDto);
     List<LessonDto> getLessonsByType(Type lessonType);
-    boolean addReview(Integer lessonId, Review review);
-    Set<Review> getReviewOnTheLesson(Integer lessonId);
+    boolean addReview(Long lessonId, Review review);
+    Set<Review> getReviewOnTheLesson(Long lessonId);
     int getRewardForLessonsByProfessor(String username);
-    boolean changeRewardByLessonId(int lessonId, int reward);
+    boolean changeRewardByLessonId(Long lessonId, int reward);
 }
