+++ final/Web/service/src/main/java/com/expexchangeservice/service/interfaces/ITokenService.java
@@ -2,7 +2,6 @@ package com.expexchangeservice.service.interfaces;
 
 import com.expexchangeservice.model.dto.UserDto;
 import com.expexchangeservice.model.entities.Token;
-import com.expexchangeservice.model.entities.User;
 import org.springframework.security.core.Authentication;
 
 import javax.servlet.ServletRequest;
@@ -15,7 +14,7 @@ public interface ITokenService {
     Authentication getAuthentication(ServletRequest servletRequest);
     Token generateToken(UserDto userDto);
     boolean checkUser(HttpServletRequest request, String username);
-    boolean checkUser(HttpServletRequest request, int userId);
+    boolean checkUser(HttpServletRequest request, Long userId);
     boolean deleteToken(HttpServletRequest request);
     String getUsernameByRequest(HttpServletRequest request);
 }
