+++ final/Web/utils/src/main/java/com/expexchangeservice/utils/security/ITokenHandler.java
@@ -1,15 +1,12 @@
 package com.expexchangeservice.utils.security;
 
-import java.util.Optional;
-
 public interface ITokenHandler {
 
     String AUTH_HEADER_NAME = "Auth-Token";
 
-    String generateToken(Integer userId);
+    String generateToken(Long userId);
 
     boolean checkToken(String token);
 
-        Integer getUserIdFromToken(String token);
-//    Optional<Integer> getUserIdFromToken(String token);
+    Long getUserIdFromToken(String token);
 }
